requirements.txt
# Core dependencies
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.11.0
email-validator>=2.0.0
sqlalchemy>=2.0.40
asyncpg>=0.30.0
aiohttp>=3.11.0
httpx>=0.28.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.10.0
python-multipart>=0.0.9
PyYAML>=6.0.0
tenacity>=8.0.0

# Database and storage
redis>=5.0.0
aioredis>=2.0.0
networkx>=3.0.0
psycopg2-binary>=2.9.0
qdrant-client>=1.13.0
faiss-cpu>=1.10.0

# WebSocket and real-time communication
websockets>=12.0

# AI and ML
numpy>=2.0.0
scipy>=1.15.0
networkx>=3.0
pandas>=2.0.0
matplotlib>=3.0.0
sentence-transformers>=4.0.0
Pillow>=10.0.0
tiktoken>=0.8.0
langchain>=0.3.25
langchain-community>=0.3.24
openai>=1.0.0

# Monitoring and logging
prometheus-client>=0.17.0
opentelemetry-api>=1.30.0
opentelemetry-sdk>=1.30.0
opentelemetry-exporter-prometheus>=0.54b1
psutil>=7.0.0

# AWS
boto3>=1.37.0

# Shipping and logistics
shippo>=3.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.26.0
pytest-cov>=4.1.0
